requirements.txt
flask-3.1.0 
flask-sqlalchemy-3.1.1
flask-restful-0.3.10
lask-jwt-extended-4.7.1